Including changes to 2009 project
I'm aiming to improve this project. The project is ... somewhat usable
in its current state. However, in order to make it fully functional,
some major changes must be made. These changes reflect what I have
learned over the past few years in college, on the job, and in my
research. In other words, I can write more modularized and maintainable
code now than when this project was created in the first place.